general update

+++ src/main/java/com/bensiegler/calendarservice/services/AuthorityService.java
@@ -5,6 +5,7 @@ import com.bensiegler.calendarservice.models.dbmodels.Authority;
 import com.bensiegler.calendarservice.repositories.AuthorityRepo;
 import com.bensiegler.calendarservice.security.AuthenticationFacade;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -20,7 +21,7 @@ public class AuthorityService {
     AuthorityRepo authorityRepo;
 
 
-    public List<Authority> getAuthorityByCalendarObjectId(String calendarObjectId) {
+    public List<Authority> getByCalendarObjectIdAndUserId(String calendarObjectId) {
         String userId = AuthenticationFacade.getUserId();
         List<Authority> authorities =  authorityRepo.findByAuthorityGrantedToAndCalendarObjectId(userId, calendarObjectId);
 
@@ -31,9 +32,19 @@ public class AuthorityService {
         return authorities;
     }
 
-    public Authority getAuthorityByCalendarObjectIdAndPowerGiven(String calendarObjectId, String powerGiven) {
+    public Authority getByCalendarObjectIdAndPowerGivenAndUserId(String calendarObjectId, String powerGiven) {
         return authorityRepo.findByPowerGivenAndAuthorityGrantedToAndCalendarObjectId(powerGiven,
                 AuthenticationFacade.getUserId(), calendarObjectId);
+    }
+
+    public List<Authority> getByAuthorityGrantedToAndPowerGiven(String powerGiven) {
+        return authorityRepo.findByAuthorityGrantedToAndPowerGiven(AuthenticationFacade.getUserId(), powerGiven);
+    }
+
+    public void deleteAllAuthoritiesForAnObject(String calendarObjectId) {
+        ArrayList<Authority> authoritiesToBeDeleted = (ArrayList<Authority>)authorityRepo.findByCalendarObjectId(calendarObjectId);
+
+        authorityRepo.delete(authoritiesToBeDeleted);
     }
 
 
@@ -53,8 +64,16 @@ public class AuthorityService {
         return newPowers;
     }
 
-    public void revokeAuthority(Authority authority) {
+    public void revokeAuthority(Authority authority) throws AuthorityException {
         authority.formatUUIDs();
+
+        Authority dbAuthority = authorityRepo.findOne(authority.getId());
+
+        //TODO problem since sharers can remove owners authorities which should not be possible
+        if(dbAuthority.getAuthorityGrantedTo().equals(dbAuthority.getActionTakenBy()) || dbAuthority.getPowerGiven().equals("OWNER")) {
+            throw new AuthorityException("You are the owner of this and cannot revoke your own authorities");
+        }
+
         if(null != authority.getId()) {
             authorityRepo.delete(authority.getId());
         }else{
@@ -73,7 +92,7 @@ public class AuthorityService {
     public void addAuthorityToPrincipal(String authority) {
         Collection<SimpleGrantedAuthority> oldAuthorities = (Collection<SimpleGrantedAuthority>)AuthenticationFacade.getAuthentication().getAuthorities();
         SimpleGrantedAuthority newAuthority = new SimpleGrantedAuthority(authority);
-        List<SimpleGrantedAuthority> updatedAuthorities = new ArrayList<SimpleGrantedAuthority>();
+        List<SimpleGrantedAuthority> updatedAuthorities = new ArrayList<>();
         updatedAuthorities.add(newAuthority);
         updatedAuthorities.addAll(oldAuthorities);
